+++ examples/get_user_details.rs
@@ -12,11 +12,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .map(|s| s.to_string())
         .collect::<Vec<_>>();
     // A blank key API key can be passed, it isn't needed in this example.
-    let manager = TradeOfferManager::builder(String::from(""), "./assets")
+    let manager = TradeOfferManager::builder(String::from(""))
         .identity_secret(String::from("secret"))
         .cookies(cookies) // Cookies can also be set using the `set_cookies` method on the manager
         .build();
-    
     // Passing in GetUserDetailsMethod::None assumes we are friends with the user.
     let user_details = manager.get_user_details(steamid, GetUserDetailsMethod::None).await?;