+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - `Error::Http` to `Error::StatusCode`.
 - `HOSTNAME` constants to private.
 - `ClassInfoCache` to use `Arc<Mutex<T>>` internally rather than requiring it to be wrapped.
+- `TradeOfferManagerBuilder` no longer requires a data directory. The data directory now defaults to the user's config directory.
 
 ## 0.3.0 (2023-10-08)